Create callbacks to detect upgrade module is up and ACK upgrade event Fix test_wpk_agent for Windows systems
